% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_root.R
\name{is_root}
\alias{is_root}
\alias{is_root.rpart}
\alias{is_root.BinaryTree}
\alias{is_root.train}
\title{Is the tree just a root node?}
\usage{
is_root(object, ...)

\method{is_root}{rpart}(object, ...)

\method{is_root}{BinaryTree}(object, ...)

\method{is_root}{train}(object, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{"rpart"} or
\code{"BinaryTree"}.}

\item{...}{Additional optional arguments. Currently ignored.}
}
\value{
Returns \code{TRUE} if \code{object} does not contain any splits and
  \code{FALSE} otherwise.
}
\description{
Tests if an \code{"rpart"} or \code{"BinaryTree"} object contains only a root
node (i.e., no splits).
}
