% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsplits.R
\name{nsplits}
\alias{nsplits}
\title{Number of splits}
\usage{
nsplits(object)
}
\arguments{
\item{object}{An object of class \code{"rpart"} or \code{"BinaryTree"}.}
}
\value{
The number of splits contained in \code{object}.
}
\description{
Compute the number of splits of a tree.
}
