% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{predict.rforest}
\alias{predict.rforest}
\title{Random forest predictions}
\usage{
\method{predict}{rforest}(object, newX, predict.all = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{"rftree"}.}

\item{newX}{Data frame or matrix of new feature values.}

\item{predict.all}{Logical indicating whether or not to return predictions
for each individual tree. Default is \code{FALSE}, which only returns the 
aggregated predictions.}

\item{...}{Additional optional arguments. (Currently ignored.)}
}
\value{
A vector of predictions. For binary outcomes coded as 0/1, the 
predictions represent Pr(Y = 1).
}
\description{
Compute predictions from an \code{"rftree"} object.
}
