% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{predict.rftree}
\alias{predict.rftree}
\title{Model predictions}
\usage{
\method{predict}{rftree}(object, newX)
}
\arguments{
\item{object}{An object of class \code{"rftree"}.}

\item{newX}{Data frame or matrix of new feature values.}
}
\value{
Has the same return type as 
\code{\link[randomForest]{predict.randomForest}}.
}
\description{
Compute predictions from an \code{"rftree"} object.
}
\keyword{internal}
