% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{rftree}
\alias{rftree}
\title{Random forest tree}
\usage{
rftree(X, y, rotate = FALSE, ...)
}
\arguments{
\item{X}{A data frame or a matrix of predictors.}

\item{y}{Numeric vector of response value. For binary outcomes, \code{y} 
should be mapped to {0, 1}. Note that multiclass outcomes are not supported.}

\item{rotate}{Logical indicating whether or not to randomly rotate the 
feature values prior to fitting the tree. Default is \code{FALSE} which
results in a traditional random forest tree.}

\item{...}{Optional arguments to be passed on to 
\code{\link[randomForest]{randomForest}}.}
}
\description{
Fits a single random forest tree.
}
\keyword{internal}
