% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{splitters}
\alias{splitters}
\alias{splitters.rpart}
\alias{splitters.BinaryTree}
\alias{splitters.train}
\title{Split variables}
\usage{
splitters(object)

\method{splitters}{rpart}(object)

\method{splitters}{BinaryTree}(object)

\method{splitters}{train}(object)
}
\arguments{
\item{object}{An object of class \code{"rpart"} or \code{"BinaryTree"}.}
}
\value{
The number of splits contained in \code{object}.
}
\description{
Extract the variables used as splitters in a tree.
}
