% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rforest.R
\name{suppressRegressionWarning}
\alias{suppressRegressionWarning}
\title{Suppress randomForest() warning message}
\usage{
suppressRegressionWarning(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\description{
Suppresses a specific \code{\link[randomForest]{randomForest}} warning 
message regarding the response having "...five or fewer unique values."
}
\keyword{internal}
